requirements.txt
numpy
Pillow
tqdm
opencv-python
scikit-image